# improve the launch performance of portable