For portable, use the unpack binary.

|launch time|cold launch|hot launch|
|-|-|-|
|old|>14s|>10s|
|new|<5s|<1s|