@Eugeny 2020-02-13
Looks good :+1:

I'll swap the target for zip to avoid users needing to have 7zip to unpack it.